Make token scheme its own module
